+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 
 - Fixed #38, notify stewards of new venue proposal.
 - Fixed #36, notify scholars of when they are added or removed to a submission.
+- Fixed #39, notify editors of a new venue proposal.
 
 ### Fixed